requirements.txt
telegram_send
GitPython
python-dotenv
